Change to strict spaces around property delimiter
Also remove trimming of values, this improve consistency of parse/emit.
Removing the value trimming also adds the ability to line-up the values.

    name:            Bob
    age:             34
    likes_ice_cream: true

But it removes previously allowed properties.

    key :value
    key:value
    key : value

The first one is ugly so ok to disallow.  The second is used in time,
ratios, namespaces, etc so make sense to disallow.  The third form is
the most legit use, that's where the tradeoff happens.